# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Morseverse",
    "description": "APIs for the Morseverse website",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/get-cloth/{category}": {
      "get": {
        "tags": [
          "Google Cloud"
        ],
        "summary": "Get Image",
        "description": "Fetch a list of images from GCS based on the provided category.\n\nArgs:\n    category (str): The category to fetch images for (e.g., \"tops\", \"bottoms\", \"overwears\").\n\nReturns:\n    List[str]: A list of image file paths for the given category.",
        "operationId": "get_image_api_v2_get_cloth__category__get",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Get Image Api V2 Get Cloth  Category  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/get-image/{image_path}": {
      "get": {
        "tags": [
          "Google Cloud"
        ],
        "summary": "Get Image",
        "description": "Fetch an image from GCS based on the provided path and return it in Base64 format.\nThe image path is relative to the GCS bucket.\n\nExample: images/project1/360image1.jpg",
        "operationId": "get_image_api_v2_get_image__image_path__get",
        "parameters": [
          {
            "name": "image_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageBase64Response"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/recommendation/{image_path}": {
      "get": {
        "tags": [
          "Google Cloud"
        ],
        "summary": "Get Recommendation",
        "operationId": "get_recommendation_api_v2_recommendation__image_path__get",
        "parameters": [
          {
            "name": "image_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Image Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/character_with_cloth/{main_character}": {
      "get": {
        "tags": [
          "Google Cloud"
        ],
        "summary": "Get Character Image",
        "operationId": "get_character_image_api_v2_character_with_cloth__main_character__get",
        "parameters": [
          {
            "name": "main_character",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "BucketPath",
              "title": "Main Character"
            },
            "description": "BucketPath"
          },
          {
            "name": "gender",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(man|woman)$",
              "title": "Gender"
            }
          },
          {
            "name": "cloth_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Path to the clothing resource",
              "title": "Cloth Path"
            },
            "description": "Path to the clothing resource"
          },
          {
            "name": "guidance_scale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 2.0,
              "title": "Guidance Scale"
            }
          },
          {
            "name": "timesteps",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Timesteps"
            }
          },
          {
            "name": "num_samples",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Num Samples"
            }
          },
          {
            "name": "seed",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": 42,
              "title": "Seed"
            }
          },
          {
            "name": "nsfw_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": true,
              "title": "Nsfw Filter"
            }
          },
          {
            "name": "cover_feet",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": false,
              "title": "Cover Feet"
            }
          },
          {
            "name": "adjust_hands",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": false,
              "title": "Adjust Hands"
            }
          },
          {
            "name": "restore_background",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": false,
              "title": "Restore Background"
            }
          },
          {
            "name": "restore_clothes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": false,
              "title": "Restore Clothes"
            }
          },
          {
            "name": "garment_photo_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "auto",
              "title": "Garment Photo Type"
            }
          },
          {
            "name": "long_top",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": false,
              "title": "Long Top"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Get Character Image Api V2 Character With Cloth  Main Character  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImageBase64Response": {
        "properties": {
          "image_base64": {
            "type": "string",
            "title": "Image Base64"
          }
        },
        "type": "object",
        "required": [
          "image_base64"
        ],
        "title": "ImageBase64Response"
      },
...
      "RecommendationList": {
        "properties": {
          "image_paths": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Image Paths"
          }
        },
        "type": "object",
        "required": [
          "image_paths"
        ],
        "title": "RecommendationList"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}